Bug fix: Bug id#3204925: ci_sockaddr_set_family incorrect if USE_IPV6 Author: Alexander Dupuy
There is a copy-paste error in the #define of this macro in include/net_io.h if USE_IPV6 is defined:

#define ci_sockaddr_set_family(addr,port) ((addr).sockaddr.ss_family=family)

the macro parameter 'port' should be 'family':

#define ci_sockaddr_set_family(addr,family) ((addr).sockaddr.ss_family=family)

Since this is a macro, unused in c_icap or c_icap_modules, this does not cause 
any compile errors, and may not even cause compile errors if used by a service 
plugin or other program using the library, if it happens to have a family 
variable. Even without a compile error, the code will be buggy.



git-svn-id: https://svn.code.sf.net/p/c-icap/code/c-icap-server/trunk/c-icap@763 950cb5d8-81e4-40ab-8cb8-c719d5138fe1