Bug fix: Allow records with only key and no value in ci_cache
There are cases where we need to store in ci_cache
records which have only keys (normaly key:value pairs stored), and
other cases where we do not need to allocate memory for value data.
A such case is the ci_cache used in dnsbl_tables module.
In such cases we should set the ci_cache_entry::val_size to 0 and we must
not call free handlers to free value data.



git-svn-id: https://svn.code.sf.net/p/c-icap/code/c-icap-server/trunk/c-icap@796 950cb5d8-81e4-40ab-8cb8-c719d5138fe1