Bug fix: The spaces are removed from ldap table definition string
This is a problem in the case an LDAP attribute (eg "cn") value contain spaces.
For example if someone uses the "cn=Directory Manager" as username the c-icap
will convert it to "cn=DirectoryManager".
What actually needed is to remove the leading and trailing spaces from all
fields of the ldap table definition string. This patch adds and uses the
str_trim function which does the job.



git-svn-id: https://svn.code.sf.net/p/c-icap/code/c-icap-server/branches/c_icap_0_1_x@798 950cb5d8-81e4-40ab-8cb8-c719d5138fe1