Bug fix: Fix ci_http_request_url (%huo fmt code) to work with transparent proxies
The ci_http_request_url return the second string exist in request line. For
transparent proxies this is only the path not a full URL.
This patch fixes the ci_http_request_url to return the hostname/path for 
these cases.
The hostname retrieved from Host header if exist. 
The protocol information (eg "http://" or "https://") can not be retrieved.





git-svn-id: https://svn.code.sf.net/p/c-icap/code/c-icap-server/trunk/c-icap@889 950cb5d8-81e4-40ab-8cb8-c719d5138fe1