Bug fix: Fix deadlocks in access log subsystem related to relog command
There are many cases where a deadlock  may occur while logs written
on access log files or server log file, specially if the rwlocks are not 
supported or they are disabled.

This patch uses one mutex per log file to avoid deadlocks and also try to
make as short as possible the lockings.



git-svn-id: https://svn.code.sf.net/p/c-icap/code/c-icap-server/trunk/c-icap@1008 950cb5d8-81e4-40ab-8cb8-c719d5138fe1