client headers reallocation patch Author: Jonatan Vela <jonatan.vela@ergon.ch>
Fixes a bug in ci_headers_addheaders():
Adding headers to an empty header structure with ci_headers_addheaders() caused
memory errors since h->headers[0] is not initialized in ci_headers_create().
 
Furthermore the patch contains optimizations for the reallocation of the
header buffers. ow, when reallocation is needed the final size of the buffer
is calculated and then reallocated (only once).




git-svn-id: https://svn.code.sf.net/p/c-icap/code/c-icap-server/trunk/c-icap@1146 950cb5d8-81e4-40ab-8cb8-c719d5138fe1