Make default interprocessing scheme the "file"
The posix semaphores has the problem that when one process crashed, does not
release/unlock the semaphores for which holds the lock.
This is may cause dead lock to c-icap server, for example in the case a crashed
process holds the locking of accept mutex: The remaining process waiting for
ever the accept mutex to be unlocked.



git-svn-id: https://svn.code.sf.net/p/c-icap/code/c-icap-server/trunk/c-icap@1177 950cb5d8-81e4-40ab-8cb8-c719d5138fe1