Use poll instead of select for polling network sockets and pipes io state
Using some handreds of c-icap worker threads is usual in large instalations.
In these cases the number of open file derscriptors (network  connections 
to icap client, logging filedescriptors, or connections to external services 
or databases etc), can exceed the 1024 witch is the default value of 
FD_SETSIZE on most systems. The behaviour of select when an fd greater than
the FD_SETSIZE is used, will  be undefined causing problems or crashes
to c-icap server.

This patch replaces the select function with the poll which does not have 
such limits.

Still there is the option in configure script to allow user use select
instead of poll if required.




git-svn-id: https://svn.code.sf.net/p/c-icap/code/c-icap-server/trunk/c-icap@1211 950cb5d8-81e4-40ab-8cb8-c719d5138fe1