c-icap crashes on shutdown or on reconfigure, because of unloaded c++ modules
C++ uses atexit handlers to run global objects destructors. This is problem
for c-icap services and modules which are build using c++:
 - The c-icap on reconfigure or shutdown, unloads services/modules, which are
   loaded as dynamic libraries and the memory is released
 - When a kid or main c-icap process shutdowns, runs the atexit handlers and
   the global objects destructors, but they are operate on released memory,
   causing segfault to the c-icap.

This patch add a mechanism to configure c-icap to not unload selected
services or modules:
  - Add the forceUnload=off option to the 'Service' and 'Module' configuration
    parameters.
  - Allow services or modules to use CI_MOD_DISABLE_FORCE_UNLOAD() macro to
    disable unloading of the service/module on reconfigure/shutdown.

This is may result on problems on service/module configuration parameters and 
the services/modules which is using the CI_MOD_DISABLE_FORCE_UNLOAD macro must
result global variables/parameters to default values on release/init service
handlers.



git-svn-id: https://svn.code.sf.net/p/c-icap/code/c-icap-server/trunk/c-icap@1262 950cb5d8-81e4-40ab-8cb8-c719d5138fe1