Add more options when build and using berkeleyDB lookup tables
The c-icap-mkbdb utility now create by default lookuptables which are using
hashes. The --btree option can be used to use btree instead.
Also the -p option can be used to adjust database page size.

The bdb lookup tables can take as parameter the cache size to use and the
number of caches to use.



git-svn-id: https://svn.code.sf.net/p/c-icap/code/c-icap-server/trunk/c-icap@1266 950cb5d8-81e4-40ab-8cb8-c719d5138fe1