Fix small memory leak while c-icap server shutdowns or reconfigures
In all places the destroy_childs_queue is used considered that the
passed "struct childs_queue" object memory released into this function.
However this is not happens. The create_childs_queue/destroy_childs_queue
consider (on their side) that the developer is responsible to allocate
and release the "struct childs_queue" object memory.

This patch fixes:
 - create_childs_queue to allocate required memory for struct childs_queue
   object
 - destroy_childs_queue to release memory of struct childs_queue object