Bug fix: Kids are not existing if the parent process is gone
The c-icap kids must be shutdown if the master process, which controls
their operation is gone for a reason, if the select is used instead of
poll. The select system call if the other end of the pipe is gone
retun a read event and the next read call will return error.
The c-icap unfortunatelly does not hande this case well.