Hide struct stat_memblock internal structure from users
Actually is not hided (and there is not a plan to fully hide it)
but implemented functions which can be uses instead of directly
accessing stat_memblock members.
The following functions are implemented:
 - ci_stat_groups_iterate : iterate over defined statistics groups and
   call a given callback function for each group
 - ci_stat_statistics_iterate : iterage over defined statistics and
   call a given callback function for each statistic
 - ci_stat_memblock_get_counter : Get a statistic counter value from its id
 - ci_stat_memblock_get_kbs : Get a statistic kbs counter value from its id

Also restructure info.c and proc_threads_queues.c to use new functions.
The STAT_INT64, STAT_KBS, STAT_GROUPS arrays are internal to stats.c
file