Atomic operations for c-icap server
This is mostly provided for c-icap internal use, to implement
various APIs, however maybe is helpfull for c-icap library users.
The ci_atomic interface provides API for 64bit integers which are mostly
used for statistics and counters and also for 128bit integers, if
they are supported by system, which can be used for kilobyte statistics
counters. In the future more types can be supported.

There is a try to allow C++ compilers use inlined atomic operations. I
am not sure if the result worth the added complexity, the tests show only
small performance drop when the non-inlined versions of the atomics
operations are used.

Currently the following functions are implemented:
   void  ci_atomic_load_u64(uint64_t *counter, uint64_t *value);
   void ci_atomic_add_u64(uint64_t *counter, uint64_t add);
   void ci_atomic_sub_u64(uint64_t *counter, uint64_t sub);
   void  ci_atomic_load_i64(int64_t *counter, int64_t *value);
   void ci_atomic_add_i64(int64_t *counter, int64_t add);
   void ci_atomic_sub_i64(int64_t *counter, int64_t sub);

If 128bit integers are supported (eg gcc) also the following functions
are implemented:
   void  ci_atomic_load_u128(unsigned __int128 *counter, unsigned __int128 *value);
   void  ci_atomic_add_u128(unsigned __int128 *counter, unsigned __int128 add);
   void  ci_atomic_sub_u128(unsigned __int128 *counter, unsigned __int128 sub);
   void  ci_atomic_load_i128(__int128 *counter, __int128 *value);
   void  ci_atomic_add_i128(__int128 *counter, __int128 add);
   void  ci_atomic_sub_i128(__int128 *counter, __int128 sub);

This functions maybe provided as inlined functions or as library calls
depending the compilers and architecture.