Handle unaccounted in statistics items stored in a pool_allocator
The c-icap is possible to start using a pool_allocator object to
allocate memory before the statistics interface is initialised.
For example allocate a ci_buffer object in master process before
the kids initialised and started.
These objects when will be freed later may result to negative
counters.

This patch try to handle these objects by adding a flag in allocated
object which says if the object is accounted in statistics when
allocated.