Do not use inlined functions to access structure/objects members
To use inlined functions or C macros to access structures/objects members
is problematic: The c-icap needs to act as a library and the services
and modules developers should not be restricted to use the same compiler
used to build the c-icap. However different compilers may use different
representation of strcures/objects in memory eg because of different
alignment.

The best practice is to never export structures/objects from the API
however the c-icap is not ready for this.

Recently I converted most of the C macros used to access structures
members to inlined functions. I had in my mind, as a second step,
to develop an easy way to implement both inlined and non-inlined
versions of the functions, using C macros however this is not easy.
We can always provide both inlined and non-inlined version of the
functions (this patch does it in some cases) however this is makes
the code and debuging a little complex without a huge gain.

This patch converts to normal functions most of the inlined functions
where the structures members are accessed directly.

The best practice is to never export structures/objects from the API
however the c-icap is not ready for this. A first step for this
is to separate include headers used internaly by c-icap from the
API headers, like other open source projects does.

The inlined functions still are used as function aliases/wrappers. Some
other functions (eg ci_req_lock_data/vi_req_unlovk_data) are implemented
both as inlined and non-inlined, where the inlined versions are used
only by c-icap server internally. Also some functions which used only
by the c-icap internally (eg ci_stat_membock_*) are implemented as inlined
functions even if the related structure members are accessed directly.