Default SSL_CTX object for TLS client connections looks broken
This is affects only TLS client connections where the caller/user
did not pass a SSL_CTX object to use. Currently this option
is unused in c-icap and tools code.
The major issue is that it is not thread safe, but also looks as
a code which is does not make any sense.