ci_vector_t changes
This patch adds new API functions for the ci_vector_t related API and
also allow library users to enable/disable the ci_vector stored items
memory alignment.
The alignment may required for numbers or structs have numbers as
members but it is not required for string vector items.
This patch by default uses alignment for ci_vector objects but
disable it for ci_str_vector_t objecs.
The new functions are:
  - ci_vector_set_align which enable/disable alignemnt for ci_Vector_t object
  - ci_vector_get2: similar to ci_vector_get but also return the size of
    stored object.