Fix segfault on reconfigure, when the forceUnload=off flag is used in Service configuration  parameter
For the services which uses the new _CI_DECLARE_SERVICE service registration
macro the c-icap allocates a new ci_service_t structure which must be released
on shutdown or on reconfigure where the dl-module is unloaded. In the case
the forceUnload=off flag is used this ci_service_t object is released but
c-icap trying to reuse it resulting to a segfault.