Dynamicaly allocated variable used after it is released
The CI_CONF.PIDFILE is passed to clear_pid function after it is released. This is
can cause unexpeected behaviour.
This variable stores the pid file. The pid file can not be changed after reconfigure,
so it is acceptable to use a local variable where the pid file name is copied.