Close listening ports on children shutdown procedure
... and also release related allocated memory.
The code which releases the listening ports on children shutdown was disabled
because closing TLS ports using the openSSL API resulted to close the port
not only for current/exiting child but for all of the running c-icap processes
(because OpenSSLuses shutdown system call to close socket).
This patch also fixes c-icap to tell to OpenSSL API to leave the socket fd closure
to the c-icap.